Revert "Online lateral lag learning (#34531)"
This reverts commit 1034dbd37c592630727948e5f958bb3898b377f8.
