+++ cereal/log.capnp
@@ -2278,21 +2278,6 @@ struct LiveTorqueParametersData {
   useParams @12 :Bool;
 }
 
-struct LiveDelayData {
-  lateralDelay @0 :Float32;
-  validBlocks @1 :Int32;
-  status @2 :Status;
-
-  lateralDelayEstimate @3 :Float32;
-  points @4 :List(Float32);
-
-  enum Status {
-    unestimated @0;
-    estimated @1;
-    invalid @2;
-  }
-}
-
 struct LiveMapDataDEPRECATED {
   speedLimitValid @0 :Bool;
   speedLimit @1 :Float32;
@@ -2523,7 +2508,6 @@ struct Event {
     gnssMeasurements @91 :GnssMeasurements;
     liveParameters @61 :LiveParametersData;
     liveTorqueParameters @94 :LiveTorqueParametersData;
-    liveDelay @146 : LiveDelayData;
     cameraOdometry @63 :CameraOdometry;
     thumbnail @66: Thumbnail;
     onroadEvents @134: List(OnroadEvent);
