+++ selfdrive/test/process_replay/process_replay.py
@@ -542,8 +542,8 @@ CONFIGS = [
   ),
   ProcessConfig(
     proc_name="paramsd",
-    pubs=["livePose", "liveCalibration", "carState", "carControl", "controlsState"],
-    subs=["liveParameters", "liveDelay"],
+    pubs=["livePose", "liveCalibration", "carState"],
+    subs=["liveParameters"],
     ignore=["logMonoTime"],
     init_callback=get_car_params_callback,
     should_recv_callback=FrequencyBasedRcvCallback("livePose"),
